Separated out FocusEditorService logic
