+++ fuse.py
@@ -9,6 +9,7 @@ from .go_to_definition import *
 from .version import VERSION
 from .log import log
 from .building import BuildManager
+from .focus_editor import FocusEditorService
 from . import build_results
 
 gFuse = None
@@ -34,19 +35,8 @@ class Fuse():
 		self.startFuseThread.daemon = True
 		self.startFuseThread.start()
 		self.buildManager = BuildManager(self.showFuseNotFound)
-
-	#TODO make module
-	def publishFocusEditorService(self):
-		self.msgManager.sendRequestAsync(
-			self.interop,
-			"PublishService",
-			{
-				"RequestNames" : ["FocusEditor"]
-			},
-			self.focusEditorServiceSuccess)
-
-	def focusEditorServiceSuccess(self, _):
-		log().info("Successfully registered FocusEditor service")
+		self.services = []
+		self.services.append(FocusEditorService(self.msgManager, self.interop))
 
 	def recv(self, msg):
 		try:
@@ -64,40 +54,10 @@ class Fuse():
 			log().error(traceback.format_exc())
 
 	def handleRequest(self, request):
-		if request.name == "FocusEditor":
-			if self.tryHandleFocusRequest(request):
+		for service in self.services:
+			if service.tryHandle(request):
 				return
 		self.msgManager.sendResponse(self.interop, request.id, "Unhandled")
-
-	def tryHandleFocusRequest(self, request):
-		if self.projectIsOpen(request.arguments["Project"]):
-			window = sublime.active_window()
-			view = window.open_file(
-				"{}:{}:{}".format(*[request.arguments[field] for field in ("File", "Line", "Column")]),
-				sublime.ENCODED_POSITION)
-			if sublime.platform() == "osx":
-				self.focusWindowOSX()
-				self.msgManager.sendResponse(self.interop, request.id, "Success")
-				return True
-			elif sublime.platform() == "windows":
-				self.msgManager.sendResponse(self.interop, request.id, "Success", {"FocusHwnd":window.hwnd()})
-				return True
-		return False
-
-	def projectIsOpen(self, project):
-		if not os.path.isfile(project):
-			return False
-		for folder in sublime.active_window().folders():
-			if project.startswith(folder):
-				return True
-		return False
-
-	def focusWindowOSX(self):
-		cmd = """
-			tell application "System Events"
-				activate application "Sublime Text"
-			end tell"""
-		subprocess.Popen(['/usr/bin/osascript', "-e", cmd], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
 	def showFuseNotFound(self):
 		error_message("Fuse could not be found.\n\nAttempted to run from: '"+getFusePathFromSettings()+"'\n\nPlease verify your Fuse installation." + self.rebootMessage())
@@ -234,7 +194,8 @@ class Fuse():
 
 	def onConnected(self):
 		self.sendHello()
-		self.publishFocusEditorService()
+		for service in self.services:
+			service.publish()
 
 	def sendHello(self):
 		log().info("Sending hello request")
